Feat: added Sol class

+++ entities/entity.cpp
@@ -1,7 +1,7 @@
 #include <entity.h>
 #include <SDL.h>
 
-void Entity::update(float mod){
+void Entity::update(double mod){
     show();
 }
